fix[DE]: Fix misplaced methods
